made showingCachedStock only a prop for SearchLoader
