+++ src/client/components/home/index.js
@@ -2,30 +2,11 @@ import { connect } from 'react-redux';
 
 import Home from './Home';
 
-function isStockLoading(state) {
-  const { search, stocks } = state;
-  const { lastSearch } = search;
-  if (lastSearch && stocks[lastSearch]) {
-    return stocks[lastSearch].isFetching;
-  }
-  return false;
-}
-
-function searchedStockSelected(state) {
-  const { search, stocks } = state;
-  const { lastSearch } = search;
-  if (lastSearch && stocks[lastSearch].symbol) {
-    return stocks[lastSearch].symbol === state.selectedStock;
-  }
-  return false;
-}
+import { isStockLoading } from '../../../utils/stateGetters';
 
 const mapStateToProps = state => {
-  const loading = isStockLoading(state);
-
   return {
-    loading,
-    showingCachedStock: loading && searchedStockSelected(state)
+    loading: isStockLoading(state)
   };
 };
